[fix] Combining tasks
