+++ ros2_tutorial/setup.py
@@ -20,9 +20,8 @@ setup(
     tests_require=['pytest'],
     entry_points={
         'console_scripts': [
-            'task1 = ros2_tutorial.task1_node:main',
-            'task2 = ros2_tutorial.task2_node:main',
-            'task3 = ros2_tutorial.task3_node:main'
+            'control = ros2_tutorial.control_node:main',
+            'sense = ros2_tutorial.sense_node:main',
         ],
     },
 )
